git access control & project-scope authentication
CF now uses project-scope authentication to
fulfill or deny git requests over HTTPS (required).

If the user does not have a DN
as an attribute of their CF account,
CF returns HTTP 401 (Unauthorized).

if the project or repo does not exist
based on its name parsed from the request,
CF returns HTTP 404 (Not Found).

If the project does not have a privilege
for the user, CF returns HTTP 403 (Forbidden).